login_data/producer/requirements.txt
paho-mqtt
Faker